backend/requirements.txt
Flask==1.1.2
google-cloud-translate==2.0.1
google-cloud-language==1.3.0
google-cloud-datastore==1.13.2
google-auth==1.20.1
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-auth-httplib2==0.0.4
google-api-python-client==1.10.0